+++ tests/test_decorators.py
@@ -7,12 +7,14 @@ from aiohttp_sqlalchemy import (
     DuplicateRequestKeyError,
     sa_decorator,
 )
-from tests.conftest import ClassBasedView, ClassHandler, function_handler
+from aiohttp_sqlalchemy.typedefs import THandler
+from tests.conftest import ClassBasedView, ClassHandler
 
 
 async def test_duplicate_request_key_error(
     mocked_request: Request,
     session: AsyncSession,
+    function_handler: THandler,
 ) -> None:
     assert mocked_request.get(SA_DEFAULT_KEY) is None
     mocked_request[SA_DEFAULT_KEY] = session
@@ -21,9 +23,13 @@ async def test_duplicate_request_key_error(
         await sa_decorator()(function_handler)(mocked_request)
 
 
-async def test_session_factory_not_found(mocked_request: Request) -> None:
+async def test_session_factory_not_found(
+    mocked_request: Request,
+    wrong_key: str,
+) -> None:
+    assert wrong_key not in mocked_request
     with pytest.raises(KeyError):
-        await sa_decorator('void')(ClassBasedView.get)(mocked_request)
+        await sa_decorator(wrong_key)(ClassBasedView.get)(mocked_request)
 
 
 async def test_decorated_class_based_view(mocked_request: Request) -> None:
@@ -39,7 +45,10 @@ async def test_decorated_class_handler(mocked_request: Request) -> None:
     assert isinstance(mocked_request.get(SA_DEFAULT_KEY), AsyncSession)
 
 
-async def test_decorated_function_handler(mocked_request: Request) -> None:
+async def test_decorated_function_handler(
+    mocked_request: Request,
+    function_handler: THandler,
+) -> None:
     assert mocked_request.get(SA_DEFAULT_KEY) is None
     await sa_decorator()(function_handler)(mocked_request)
     assert isinstance(mocked_request.get(SA_DEFAULT_KEY), AsyncSession)